TypedHandler: extract delegate iteration into a method
intellij-monorepo: 9b5198d832baac12d4522fdae7232c8b142b26bf

+++ intellij-community/platform/lang-impl/src/com/intellij/codeInsight/editorActions/TypedHandler.java
@@ -49,6 +49,7 @@ import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.*;
+import java.util.function.Function;
 
 public class TypedHandler extends TypedActionHandlerBase {
   private static final Set<Character> COMPLEX_CHARS =
@@ -154,21 +155,10 @@ public class TypedHandler extends TypedActionHandlerBase {
       }
 
       Editor editor = injectedEditorIfCharTypedIsSignificant(charTyped, originalEditor, originalFile);
-      PsiFile file = editor == originalEditor ? originalFile : psiDocumentManager.getPsiFile(editor.getDocument());
-
-
-      final List<TypedHandlerDelegate> delegates = TypedHandlerDelegate.EP_NAME.getExtensionList();
+      PsiFile file = editor == originalEditor ? originalFile : Objects.requireNonNull(psiDocumentManager.getPsiFile(editor.getDocument()));
 
       if (caret == originalEditor.getCaretModel().getPrimaryCaret()) {
-        boolean handled = false;
-        for (TypedHandlerDelegate delegate : delegates) {
-          final TypedHandlerDelegate.Result result = delegate.checkAutoPopup(charTyped, project, editor, file);
-          handled = result == TypedHandlerDelegate.Result.STOP;
-          if (result != TypedHandlerDelegate.Result.CONTINUE) {
-            break;
-          }
-        }
-
+        boolean handled = callDelegates(delegate -> delegate.checkAutoPopup(charTyped, project, editor, file));
         if (!handled) {
           autoPopupCompletion(editor, charTyped, project, file);
           autoPopupParameterInfo(editor, charTyped, project, file);
@@ -180,28 +170,16 @@ public class TypedHandler extends TypedActionHandlerBase {
         return;
       }
 
-      for (TypedHandlerDelegate delegate : delegates) {
-        final TypedHandlerDelegate.Result result = delegate.beforeSelectionRemoved(charTyped, project, editor, file);
-        if (result == TypedHandlerDelegate.Result.STOP) {
-          return;
-        }
-        if (result == TypedHandlerDelegate.Result.DEFAULT) {
-          break;
-        }
+      if (callDelegates(delegate -> delegate.beforeSelectionRemoved(charTyped, project, editor, file))) {
+        return;
       }
 
       EditorModificationUtil.deleteSelectedText(editor);
 
       FileType fileType = getFileType(file, editor);
 
-      for (TypedHandlerDelegate delegate : delegates) {
-        final TypedHandlerDelegate.Result result = delegate.beforeCharTyped(charTyped, project, editor, file, fileType);
-        if (result == TypedHandlerDelegate.Result.STOP) {
-          return;
-        }
-        if (result == TypedHandlerDelegate.Result.DEFAULT) {
-          break;
-        }
+      if (callDelegates(delegate -> delegate.beforeCharTyped(charTyped, project, editor, file, fileType))) {
+        return;
       }
 
       if (')' == charTyped || ']' == charTyped || '}' == charTyped) {
@@ -233,15 +211,10 @@ public class TypedHandler extends TypedActionHandlerBase {
         indentClosingParenth(project, editor);
       }
 
-      for (TypedHandlerDelegate delegate : delegates) {
-        final TypedHandlerDelegate.Result result = delegate.charTyped(charTyped, project, editor, file);
-        if (result == TypedHandlerDelegate.Result.STOP) {
-          return;
-        }
-        if (result == TypedHandlerDelegate.Result.DEFAULT) {
-          break;
-        }
+      if (callDelegates(delegate -> delegate.charTyped(charTyped, project, editor, file))) {
+        return;
       }
+
       if ('{' == charTyped) {
         indentOpenedBrace(project, editor);
       }
@@ -249,6 +222,20 @@ public class TypedHandler extends TypedActionHandlerBase {
         indentOpenedParenth(project, editor);
       }
     });
+  }
+
+  // returns true if any delegate requested a STOP
+  private static boolean callDelegates(Function<TypedHandlerDelegate, TypedHandlerDelegate.Result> action) {
+    for (TypedHandlerDelegate delegate : TypedHandlerDelegate.EP_NAME.getExtensionList()) {
+      TypedHandlerDelegate.Result result = action.apply(delegate);
+      if (result == TypedHandlerDelegate.Result.STOP) {
+        return true;
+      }
+      if (result == TypedHandlerDelegate.Result.DEFAULT) {
+        break;
+      }
+    }
+    return false;
   }
 
   private static void type(Editor editor, char charTyped) {
